add belief_table
