+++ tests/teaching/test_belief.py
