+++ tests/teaching/test_dtmc_evaluation.py
