+++ tests/teaching/test_lovejoy.py
